feat(dto): 메뉴 그룹 mockito Happy case 작성
